Create TodoList.js
The event obj (e), contains the string of text the user inputs. This can be accessed in the method using event.target.value. 
This data is then set as the value in  state (task) & then rendered as the value in the input box on the page with value={task}. 
The component is said to be the 'single source of truth' regarding the input data. 
This is known as controlled input (or controlled forms) as React can control the internal state for certain elements like input and textarea, which makes them controlled components!